<227> Refactor project related url to project folder
