+++ scrumate/settings.py
@@ -119,7 +119,7 @@ USE_L10N = True
 
 USE_TZ = True
 
-PAGE_SIZE = 10
+PAGE_SIZE = 20
 
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/2.1/howto/static-files/
